Move misc utils in a separate header
